nit: abstract prov server
